Refactor: Home 페이지 컴포넌트 최적화, 기능 분리
