+++ src/components/home/HomeDescription.tsx
@@ -1,0 +1,10 @@
+const HomeDescription = () => {
+    return (
+        <div className="text-gray-200 text-sm select-none absolute bottom-5 right-10 animate-blur">
+            <p>해당 웹페이지는 현재 모바일을 지원하지 않습니다.</p>
+            <p>This webpage does not currently support mobile.</p>
+        </div>
+    );
+};
+
+export default HomeDescription;
